// types.ts
export interface Coupon {
    id: string;
    code: string;
    discountType: 'Fixed' | 'Percentage';
    discountValue: number;
    expiresAt?: Date;
    isActive: boolean;
    usageLimit: number; // 0 for unlimited
    timesUsed: number;
    createdAt: Date;
}
// ...
export enum NotificationType {
    SUCCESS = 'success',
    ERROR = 'error',
    INFO = 'info',
}